// src/interfaces/affliate.interface.ts
export interface Affliate {
  _id: string;
  full_name: string;
  email: string;
  phone: string;
  companyName: string;
  website: string;
  industry: string;
  document: string;
  experience: string;
  promotionMethod: string;
  comments: string;
  status: string;
  description: string;
  affliate_id: string;
  createdAt: string;
  address: string;
  documents: { url: string; publicKey: string }[];
}